Minor fixes, gpg root dir stuff, diagnostics script
git-svn-id: https://svn.surfnet.nl/svn/ids/tunnel/trunk@1896 6e3b1d27-7b68-4cde-b5d6-ae3f64e5b54f

+++ scripts/rrd_serverinfo.pl
@@ -448,5 +448,7 @@ sub RegistDB {
     $execute_result = $sth->execute();
   }
   $dbh = "";
-  `rm $c_imgdir/$file`;
+  if (-e "$c_imgdir/$file") {
+    `rm $c_imgdir/$file`;
+  }
 }
